Rename Vector2 now that we're using modules
